+++ src/lib/components/general/icons/decorative/starfilled.svelte
@@ -1,0 +1,5 @@
+<svg width="111" height="110" viewBox="0 0 111 110" fill="none" xmlns="http://www.w3.org/2000/svg">
+	<path
+		d="M55.4999 0C56.5738 29.9196 80.5804 53.9261 110.5 55C80.5804 56.0739 56.5738 80.0804 55.4999 110C54.4261 80.0804 30.4195 56.0739 0.499939 55C30.4195 53.9261 54.4261 29.9196 55.4999 0Z"
+	/>
+</svg>
