+++ src/lib/components/general/icons/index.ts
@@ -54,6 +54,7 @@ import FacebookIcon from './FacebookIcon.svelte';
 import ApplicationsIcon from './ApplicationsIcon.svelte';
 import CircledDoodle from './doodles/CircledDoodle.svelte';
 import FilterIcon from './FilterIcon.svelte';
+import StarFilledIcon from './decorative/starfilled.svelte';
 
 export {
 	HomeIcon,
@@ -111,5 +112,6 @@ export {
 	DiscordIcon,
 	CircledDoodle,
 	ApplicationsIcon,
-	FilterIcon
+	FilterIcon,
+	StarFilledIcon
 };
