+++ src/lib/components/layouts/MainPage.svelte
@@ -45,7 +45,6 @@
 		overflow: scroll;
 		flex: 1;
 		height: 100%;
-		padding: 0.5rem;
 	}
 
 	#side-content {
