--- src/lib/static/templates/developer-posting-template.md
@@ -1,47 +1,0 @@
-# Developer Posting Template
-
-> Applications are reviewed on a rolling basis. Please apply as soon as possible.
-
-### Summary
-
-Are you ready to dive into an exciting development adventure? Join the extraordinary team at UBC Launch Pad and be part of a 4-8 month journey filled with innovation and creativity! As a developer, you will collaborate with talented designers and fellow developers to bring a unique project to life. Guided by an experienced tech lead, you'll have the opportunity to unleash your full potential and make a real impact.
-
-Discover more about us on our captivating Website.
-
----
-
-### What You'll Be Doing
-
-- Unleash your coding prowess by developing cutting-edge backend API features for our internal portal.
-  - Harness the power of AWS to deploy and implement our robust backend infrastructure.
-
----
-
-### What You Get
-
-- _[Project specific]_
-- Unlock the doors to an incredible documentation website, equipped with all the resources you need to conquer challenges and achieve success.
-- Embrace professional development practices and sharpen your skills under the guidance of industry experts.
-- Experience the thrill of the full project lifecycle, from ideation to execution, and witness your creation come to life.
-- Showcase your problem-solving prowess as you collaborate with your exceptional team to build something meaningful.
-- Forge lasting bonds and create lifelong friendships with the brilliant minds at Launch Pad 💫
-
----
-
-### Time Commitment
-
-- Gear up for club meetings held every other Saturday from 12pm-3pm. The first hour is crucial, as it sets the stage for collaboration and inspiration. After that, it's time to unwind, work together, and relish a delightful lunch. We strongly encourage you to stay for the entire duration, as the magic happens during these valuable moments.
-- Get ready for bi-weekly sub-team meetings, lasting 1 hour, meticulously scheduled to accommodate everyone's availability. These meetings, alternating with the club-wide Saturday gatherings, foster deeper connections within your dedicated sub-team.
-- Dedicate approximately 5 hours per week to conquering tasks and taking your project to new heights.
-
----
-
-### About You
-
-- Buckle up and be prepared to attend all meetings with unwavering commitment and dedication.
-- Fuel your passion for learning, experimentation, and active participation, as we embrace your unique perspective and contributions.
-- At Launch Pad, we celebrate diversity, skill levels, and experiences. Whether you're a beginner embarking on an exciting journey or a seasoned developer with a portfolio full of awe-inspiring projects, you'll find your place here.
-
-##### Bonus
-
-- Have you dabbled in or possess a burning interest in []? Let us know! We're excited to hear about your diverse experiences and how they can contribute to our dynamic team.
