--- src/routes/(portal)/projects/directory/[projectName]/+page.ts
@@ -1,14 +1,0 @@
-export async function load({ params }) {
-	const post = await import(`../${params.projectName}.md`);
-	const { title, date, id } = post.metadata;
-	const content = post.default;
-
-	return {
-		content,
-		title,
-		date,
-		id
-	};
-}
-
-export const prerender = false;
