--- src/routes/docs/README.md
@@ -1,94 +1,0 @@
-# Club Handbook
-
-Welcome to the [UBC Launch Pad](https://ubclaunchpad.com/) handbook! Launch Pad-wide documentation is hosted in these sections - if you're not sure about something, this is the first place to check. The easiest way to find something is to run a search in the navigation bar!
-
-Our goals with this handbook are to:
-
-- 📝 Create a comprehensive set of documentation that anyone can leverage to participate in and run Launch Pad - with written processes available, everyone has an equal opportunity to understand what we do and participate, rather than just those "in the know".
-- 🧠 Accumulate resources and lessons learned so that future generations of Launch Pad can focus on iterating and improving, rather than figuring out how to run the club from scratch every year.
-- 💪 Empower members to find information they need independently, and hopefully enable others to learn from our work.
-- 💬 Create a culture where people are willing to share and learn from others' experiences.
-
-For this to work, **we need your help to keep the handbook up to date**! That means proactively making changes when you see something incorrect, outdated, or that you disagree with, and participating in [pull requests to change the handbook](https://github.com/ubclaunchpad/docs/pulls). To learn more about how you can contribute, take a look at our [Contribution Guide](/CONTRIBUTING.md), or reach out for help in [`#ask-for-help`](https://ubclaunchpad.slack.com/messages/CJXM08QBB)!
-
-## 🔖 Manifesto
-
-Our [**Manifesto**](manifesto.md) outlines our vision, what we do, and our code of conduct.
-
-## 🚀 Onboarding
-
-These pages helps you hit the ground running once you've joined us!
-
-- [**for Everyone**](./onboarding/everyone.md) - read this one first!
-- [**for Design**](./onboarding/design.md)
-- [**for Strategy**](./onboarding/strategy.md)
-- [**for Leads**](./onboarding/leads.md)
-
-## ⚙️ Tools
-
-We use a variety of technology and tools at Launch Pad - these sections help you get started on what we use!
-
-- [**Slack**](./tools/slack.md): Best practices, channels, bots, and general information about our Slack workspace! Slack is where most Launch Pad discussions and announcements happen.
-- [**GitHub**](./tools/github.md): Guidance on getting started with our GitHub organization! GitHub is where we host, document, and manage the code in all our projects.
-- [**Google Drive**](./tools/drive.md): How to use Launch Pad's shared Google Drive folders.
-- [**Deployment**](./tools/deployment.md): Launch Pad's recommendations regarding deployment and getting your project out to the world.
-- [**Analytics**](./tools/analytics.md): Launch Pad's recommendations regarding setting up analytics to see how users interact with your projects.
-- [**Social Media**](./tools/social-media.md): Documentation about our social media accounts and how to use them to help showcase our work!
-- [**Email**](./tools/email.md): Our email accounts and how to use them.
-- [**Other Tools**](../resources/tools.md): Our Resources section has a page on general tools you can leverage to help you build awesome stuff!
-
-## 👨‍💼 Project Management
-
-These documents compile hard-won advice about the best way to run (and participate in) a Launch Pad team and build great software. This isn't just geared towards team leads - it should be useful for everyone!
-
-::: tip
-These are **guidelines**, not strict rules! Teams should adjust these processes to the way they work. Open a pull request if there's something you think should be updated or changed!
-:::
-
-- [**Project Scope**](./project-management/scope.md): Tips for planning out your project's overall scope and getting an initial idea laid out.
-- [**Sprint Planning**](./project-management/sprints.md): Guidelines and advice for how to plan out work with your team to drive forward progress on your project!
-- [**Repository Management**](./project-management/repositories.md): Guidelines for how to manage your GitHub repository and enforce coding standards. Remember: your repository is your main outlet to the outside world - and this outside world includes potential employers! Make it informative, organized, and pretty!
-
-## 💼 Strategy
-
-These pages document our efforts in improving Launch Pad and our reach. While the [Strategy team](/handbook/onboarding/strategy.md) will own some of content and processes described in these pages, these documents still involve everyone in Launch Pad!
-
-- [**Recurring Processes**](./strategy/recurring-processes.md): Administrative processes that we run every year.
-- [**Objectives**](./strategy/objectives.md): Launch Pad's yearly objectives, goals, and our learnings.
-- [**Sponsors and Partners**](./strategy/sponsors-and-partners.md): How we work with sponsors and partners!
-
-## 🎨 Design
-
-These pages document Launch Pad design standards. For more about how the design team works, refer to the [Design team](/handbook/onboarding/strategy.md) onboarding.
-
-- [Branding](./design/branding.md): How to represent Launch Pad!
-
-## 🗳️ Recruitment
-
-Launch Pad recruitment documentation and material lives in these pages!
-
-- [**Overview**](/handbook/recruitment/overview.md) goes over the why and how of recruitment at Launch Pad.
-- **Role descriptions**: These pages describe roles that Launch Pad recruits for. They should be written such that we can send each page individually to prospective candidates to help them get an idea of what each role entails!
-  - [Developer](./recruitment/developer.md)
-  - [Design](./recruitment/designer.md)
-  - [Strategy](./recruitment/strategy.md)
-  - [Design Lead](./recruitment/design-lead.md)
-  - [Tech Lead](./recruitment/tech-lead.md)
-
-## 📦 Official Repositories
-
-There are a number of other important repositories that provide places for documentation and discussion. Note that these are all private to the `ubclaunchpad` GitHub organization; if you have not been added to the organization, talk to your Tech Lead or one of the Presidents in [`#ask-leads`](https://ubclaunchpad.slack.com/messages/CK935RD3Q/).
-
-- [**Ideas**](https://github.com/ubclaunchpad/ideas): The goal of this repository is to be a place where anyone in Launch Pad can submit ideas for anything - how to improve Launch Pad, potential projects, event ideas... anything you want!
-- [**Alumni Directory**](https://github.com/ubclaunchpad/alumni-directory): An internal directory with past Launch Pad members and what they are doing now.
-- [**ubclaunchpad.com**](https://github.com/ubclaunchpad/ubclaunchpad.com): This repository houses the source code for our website, as well as documentation on how to use it, update its content, and make your own contributions!
-- [**Docs**](https://github.com/ubclaunchpad/ideas): The repository that hosts the content on this website! Changes to `docs.ubclaunchpad.com` happen here.
-
-Some teams also have dedicated repositories:
-
-- [**Strategy**](https://github.com/ubclaunchpad/strategy): The main repo for the strategy team, containing sponsorship-related discussion.
-- [**Design**](https://github.com/ubclaunchpad/design): The main repo for the design team, containing Launch Pad graphical assets and design-related discussion.
-- [**Tech Leads**](https://github.com/ubclaunchpad/tech-leads): A repo for Tech Lead-specific content, such as meeting notes. Note that this repo is only visible to Tech Leads.
-- [**Exec**](https://github.com/ubclaunchpad/exec): Repo containing details about managing Launch Pad, such as our accounts.
-
-Note that links to most of these repositories are available through the handy "Quick Links" dropdown in this website's navbar!
