--- src/routes/docs/design/Overview.md
@@ -1,76 +1,0 @@
-<script>
-import Tip from "$lib/components/blogComponents/Tip.svelte";
-import HeadingIcon from "$lib/components/blogComponents/HeadingIcon.svelte";
-</script>
-
-# Overview
-
-<Tip>
-At Launch Pad, designers play a crucial role to our development. As a designer, you will be working closely with the developers to create a user-friendly and intuitive interface for our users. You will be responsible for the design of your team's project
-</Tip>
-
-## What you will be doing
-
-- As part of a project, you will be designing amongst developers and other designers. This inlcudes but not limited to wireframing, prototyping, and creating the final design.
-- In a designer cohort, you will be working with other designers to create and update design systems for Launch Pad. This will be used by all designers and developers to ensure consistency across all projects.
-- You will be working with the developers to ensure that the design is implemented correctly and to make any changes if necessary.
-- Take up independent design projects with other designers to create icon sets, illustrations, and other design assets for open source use, or commercial use.
-
-## What you will learn
-
-As a member of the Launch Pad Club, you will have the opportunity to learn and enhance various design skills, including:
-
-1. User Experience (UX) Design: You will gain practical experience in creating user-centered designs that prioritize usability and enhance the overall user experience.
-2. User Interface (UI) Design: You will learn how to design visually appealing interfaces that align with the project's goals and target audience.
-3. Design Thinking: You will be exposed to the design thinking process, which involves empathizing with users, defining problem areas, ideating creative solutions, prototyping, and iterating based on user feedback.
-4. Wireframing and Prototyping: You will become proficient in using design tools to create wireframes and interactive prototypes that effectively communicate design concepts and interactions.
-5. Collaboration and Communication: Working closely with developers, you will learn effective collaboration and communication skills to bridge the gap between design and development teams.
-
-## What we expect from you
-
-As a designer in the Launch Pad, we expect you to:
-
-1. Have a strong passion for design and a keen eye for aesthetics.
-2. Possess a solid understanding of design principles and best practices.
-3. Be proficient in design tools such as Sketch, Figma, or similar tools.
-4. Show a willingness to learn and adapt to new design trends and technologies.
-5. Be able to work effectively in a team environment and communicate your design ideas clearly.
-
-## What you can expect from us
-
-As a member of the Launch Pad, you can expect the following:
-
-### Design Cohort
-
-You will be part of a supportive and collaborative design cohort, where you can share ideas, seek feedback, and learn from fellow designers. The cohort will provide a platform for knowledge sharing and skill development.
-
-### Design Projects
-
-You will have the opportunity to work on real-world design projects within the club. These projects may include designing web or mobile applications, creating marketing materials, or developing branding elements. This hands-on experience will help you build a strong portfolio.
-
-### Design Reviews
-
-Regular design reviews will be conducted by experienced designers and developers within the club. These reviews will provide constructive feedback to help you improve your designs and learn from experienced professionals.
-
-### Design Systems
-
-You will contribute to the development and maintenance of design systems within Launch Pad. These design systems will ensure consistency across projects and enable efficient collaboration between designers and developers.
-
-### Design Tools
-
-You will have access to industry-standard design tools and resources. The club will provide guidance on using these tools effectively and help you stay updated with the latest design technologies and trends.
-
-### Perks and Benefits <HeadingIcon/>
-
-You hope to have access to various perks and benefits offered by the club, including free software licenses, discounts on design tools, and access to design-related events and conferences. We are actively trying to partner with companies to provide more perks to our members.
-
-## How to get help
-
-If you need any assistance or have questions, you can:
-
-1. Reach out to the project leads or mentors assigned to your team for guidance and support.
-2. Attend regular club meetings and design cohort sessions to discuss challenges and seek help from fellow designers.
-3. Utilize online resources, tutorials, and documentation recommended by the club to enhance your design skills.
-4. Engage with the club's online community or communication channels to
-
-interact with other members and ask for assistance. 5. Collaborate with your team members and utilize their expertise and perspectives to overcome design-related obstacles.
