--- src/routes/docs/onboarding/Getting-started.md
@@ -1,1 +1,0 @@
-# Getting Started
