+++ src/routes/portal/applications/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import InProgress from '$lib/components/blocks/InProgress.svelte';
+</script>
+
+<InProgress title="Main page" description="This page is still in progress." />
