+++ src/routes/portal/recruitment/postings/[id]/+page.ts
@@ -1,0 +1,7 @@
+export const load = async ({ params }) => {
+	return {
+		postingId: params.id
+	};
+};
+
+// export const ssr = false;
