+++ src/routes/variables.scss
@@ -7,6 +7,8 @@ html[data-theme='dark'] {
 	--color-bg-2: #222226;
 	--color-bg-3: #303035;
 
+	--color-bg-0-faded: #1c1c1e9f;
+
 	--color-bg-primary: #6b62b1;
 	--color-bg-primary-faded: #6b62b171;
