+++ svelte.config.js
@@ -12,15 +12,14 @@ const config = {
 	extensions: ['.svelte', '.md'],
 
 	kit: {
-        adapter: adapter({
-            // See below for an explanation of these options
-            routes: {
-                include: ['/*'],
-                exclude: ['<all>']
-            }
-        })
-    },
-	
+		adapter: adapter({
+			// See below for an explanation of these options
+			routes: {
+				include: ['/*'],
+				exclude: ['<all>']
+			}
+		})
+	},
 
 	preprocess: [
 		sveltePreprocess({
